requirements.txt
annotated-types==0.7.0
asgiref==3.7.2
crispy-bootstrap5==0.7
Django==4.2.7
django-crispy-forms==2.1
djangorestframework==3.15.2
MarkupSafe==3.0.2
Pillow==10.1.0
psycopg2-binary==2.9.9
pydantic==2.10.6
pydantic_core==2.27.2
python-dotenv==1.0.0
sqlparse==0.4.4
typing_extensions==4.12.2
tzdata==2023.3
